refactor : Lottos 로또들의 결과 처리 기능
